requirements.txt
# Editable Git install with no remote (bitpaper==1.0.0)
-e /Users/johanlescure/BitPaper
coverage==7.9.2
exceptiongroup==1.3.0
iniconfig==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
packaging==25.0
Pillow==10.1.0
pluggy==1.6.0
psutil==7.0.0
Pygments==2.19.2
pytest==8.4.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
reportlab==4.0.7
tomli==2.2.1
typing_extensions==4.14.1
reedsolo==1.7.0
cryptography==41.0.7
qrcode==7.4.2
opencv-python==4.8.1.78

